Prototype of a cutlist joiner
This still doesn't work, at least for this specific instance of a 25fps
MP4 getting joined to a 30fps recording. But otherwise, this might be
the future even for cutting up recordings.

Allowing for multiple videos to be cut up/joined/rearranged will need
more intelligence in the editor, but that comes later